+++ kapitan/cli.py
@@ -26,6 +26,7 @@ import os
 import sys
 
 import yaml
+from kapitan import cached
 from kapitan.errors import KapitanError, RefHashMismatchError
 from kapitan.initialiser import initialise_skeleton
 from kapitan.lint import start_lint
@@ -36,8 +37,8 @@ from kapitan.refs.secrets.vault import VaultSecret
 from kapitan.refs.secrets.gpg import GPGSecret, lookup_fingerprints
 from kapitan.resources import (inventory_reclass, resource_callbacks,
                                search_imports)
-from kapitan.targets import compile_targets
-from kapitan.inputs.jinja2_filters import default_jinja2_filters_path
+from kapitan.targets import compile_targets, schema_validate_compiled
+from kapitan.inputs.jinja2_filters import DEFAULT_JINJA2_FILTERS_PATH
 from kapitan.utils import (PrettyDumper, check_version, deep_get, fatal_error,
                            flatten_dict, from_dot_kapitan, jsonnet_file,
                            search_target_token_paths, searchvar)
@@ -76,7 +77,7 @@ def main():
                                 help='set search paths, default is ["."]')
     compile_parser.add_argument('--jinja2-filters', '-J2F', type=str,
                                 default=from_dot_kapitan('compile', 'jinja2-filters',
-                                default_jinja2_filters_path),
+                                                         DEFAULT_JINJA2_FILTERS_PATH),
                                 metavar='FPATH',
                                 help='load custom jinja2 filters from any file, default is to put\
                                 them inside lib/jinja2_filters.py')
@@ -96,6 +97,9 @@ def main():
     compile_parser.add_argument('--fetch',
                                 help='fetches external dependencies', action='store_true',
                                 default=from_dot_kapitan('compile', 'fetch', False))
+    compile_parser.add_argument('--validate',
+                                help='validate compile output against schemas as specified in inventory',
+                                action='store_true', default=from_dot_kapitan('compile', 'validate', False))
     compile_parser.add_argument('--targets', '-t', help='targets to compile, default is all',
                                 type=str, nargs='+',
                                 default=from_dot_kapitan('compile', 'targets', []),
@@ -129,6 +133,9 @@ def main():
                                 help='ignore the version from .kapitan',
                                 action='store_true',
                                 default=from_dot_kapitan('compile', 'ignore-version-check', False))
+    compile_parser.add_argument('--schemas-path',
+                                 default=from_dot_kapitan('validate', 'schemas-path', './schemas'),
+                                 help='set schema cache path, default is "./schemas"')
 
     inventory_parser = subparser.add_parser('inventory', help='show inventory')
     inventory_parser.add_argument('--target-name', '-t',
@@ -226,11 +233,32 @@ def main():
                              default=from_dot_kapitan('lint', 'inventory-path', './inventory'),
                              help='set inventory path, default is "./inventory"')
 
-    init_parser = subparser.add_parser('init', help='initialize a directory with the recommended kapitan project skeleton.')
+    init_parser = subparser.add_parser('init',
+                                       help='initialize a directory with the recommended kapitan project skeleton.')
     init_parser.add_argument('--directory',
                              default=from_dot_kapitan('init', 'directory', '.'),
-                             help='set path, in which to generate the project skeleton, assumes directory already exists. default is "./"')
+                             help='set path, in which to generate the project skeleton,'
+                             'assumes directory already exists. default is "./"')
 
+    validate_parser = subparser.add_parser('validate',
+                                           help='validates the compile output against schemas as specified in inventory')
+    validate_parser.add_argument('--compiled-path',
+                                 default=from_dot_kapitan('compile', 'compiled-path', './compiled'),
+                                 help='set compiled path, default is "./compiled')
+    validate_parser.add_argument('--inventory-path',
+                                 default=from_dot_kapitan('compile', 'inventory-path', './inventory'),
+                                 help='set inventory path, default is "./inventory"')
+    validate_parser.add_argument('--targets', '-t', help='targets to validate, default is all',
+                                 type=str, nargs='+',
+                                 default=from_dot_kapitan('compile', 'targets', []),
+                                 metavar='TARGET'),
+    validate_parser.add_argument('--schemas-path',
+                                 default=from_dot_kapitan('validate', 'schemas-path', './schemas'),
+                                 help='set schema cache path, default is "./schemas"')
+    validate_parser.add_argument('--parallelism', '-p', type=int,
+                                 default=from_dot_kapitan('validate', 'parallelism', 4),
+                                 metavar='INT',
+                                 help='Number of concurrent validate processes, default is 4')
     args = parser.parse_args()
 
     logger.debug('Running with args: %s', args)
@@ -240,6 +268,9 @@ def main():
     except IndexError:
         parser.print_help()
         sys.exit(1)
+
+    # cache args where key is subcommand
+    cached.args[sys.argv[1]] = args
 
     if hasattr(args, 'verbose') and args.verbose:
         logging.basicConfig(level=logging.DEBUG,
@@ -276,12 +307,17 @@ def main():
             check_version()
 
         ref_controller = RefController(args.secrets_path)
+        # cache controller for use in reveal_maybe jinja2 filter
+        cached.ref_controller_obj = ref_controller
+        cached.revealer_obj = Revealer(ref_controller)
 
         compile_targets(args.inventory_path, search_paths, args.output_path,
                         args.parallelism, args.targets, ref_controller,
                         prune=(args.prune), indent=args.indent, reveal=args.reveal,
                         cache=args.cache, cache_paths=args.cache_paths,
-                        fetch_dependencies=args.fetch, jinja2_filters=args.jinja2_filters)
+                        fetch_dependencies=args.fetch, validate=args.validate,
+                        schemas_path=args.schemas_path,
+                        jinja2_filters=args.jinja2_filters)
 
     elif cmd == 'inventory':
         if args.pattern and args.target_name == '':
@@ -323,6 +359,10 @@ def main():
             secret_update(args, ref_controller)
         elif args.update_targets or args.validate_targets:
             secret_update_validate(args, ref_controller)
+
+    elif cmd == 'validate':
+        schema_validate_compiled(args.targets, inventory_path=args.inventory_path, compiled_path=args.compiled_path,
+                                 schema_cache_path=args.schemas_path, parallel=args.parallelism)
 
 
 def secret_write(args, ref_controller):