+++ scripts/after_deploy.sh
@@ -7,8 +7,18 @@ notify_slack() {
     return 0
   fi
 
-  curl -s --retry 3 --retry-delay 3 -X POST --data-urlencode 'payload={"text": "'"$1"'"}' $SLACK_WEBHOOK > /dev/null
+  curl -s --retry 3 --retry-delay 3 -X POST --data-urlencode 'payload={"text": "'"$1"'"}' ${SLACK_WEBHOOK} > /dev/null
+}
+
+notify_hangouts() {
+  if [ -z "$HANGOUTS_WEBHOOK" ]; then
+    return 0
+  fi
+
+  curl -s --retry 3 --retry-delay 3 -H 'Content-Type: application/json' -X POST -d '{"text": "'"$1"'"}' ${HANGOUTS_WEBHOOK} > /dev/null
 }
 
 export LATEST_TAG=$(git describe --abbrev=0 --tags)
-notify_slack "Succesfully deployed ${LATEST_TAG} on Kapitan. https://github.com/deepmind/kapitan/releases/tag/${LATEST_TAG}"
+MSG="Succesfully deployed ${LATEST_TAG} on Kapitan. https://github.com/deepmind/kapitan/releases/tag/${LATEST_TAG}"
+notify_slack "${MSG}"
+notify_hangouts "${MSG}"