+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+## 0.24.0:
+- Add standalone binary to github releases
+- Upgrade some packages in requirements.txt (#344)
+- Upgrade kapp and kbld to v0.11.0 (#344)
+- Fix dependency manager's behavior for files to unpack (#342)
+- creating standalone binary for kapitan (#323)
+- Add reveal_maybe custom jinja2 filter (#332)
+- Add support to import Helm input type (#307)
+- Make boto3 requirement more flexible (#320)
+- Add kubernetes manifest validation (#317)
+- Improve json schema validation error display for inventory (#318)
+- Multi-document yaml outputs (#308)
+- Add external dependency management (#304)
+- Fix bug in resources.py when passing config to reclass (#296)
+- Fix requests version to circumvent urllib3 version conflict (#300)
+- Upgrade to jsonnet 0.13.0 (#309)
+- Add kapp and kbld to kapitan-ci (#314)
+- Implement secret sub-variables (#282)
+
+## 0.24.0-rc.6:
+- Testing github release of standalone binary
+
+## 0.24.0-rc.4:
+- Upgrade some packages in requirements.txt (#344)
+- Upgrade kapp and kbld to v0.11.0 (#344)
+- Fix dependency manager's behavior for files to unpack (#342)
+- creating standalone binary for kapitan (#323)
+
+## 0.24.0-rc.3:
+- Add reveal_maybe custom jinja2 filter (#332)
+- Add support to import Helm input type (#307)
+
+## 0.24.0-rc.2:
+- Make boto3 requirement more flexible (#320)
+- Add kubernetes manifest validation (#317)
+- Improve json schema validation error display for inventory (#318)
+- Multi-document yaml outputs (#308)
+
 ## 0.24.0-rc.1:
 - Add external dependency management (#304)
 
@@ -7,7 +45,6 @@
 - Upgrade to jsonnet 0.13.0 (#309)
 - Add kapp and kbld to kapitan-ci (#314)
 - Implement secret sub-variables (#282)
-
 
 ## 0.23.1:
 - Fix pypi package by including requirements.txt (#287)